requirements.txt
flask
flask-cors
langchain
langchain-chroma
langchain-huggingface
langchain-community
sentence-transformers
chromadb
pymupdf
requests
pypdf
python-docx
google-genai